requirements.txt
###### Peter's Starter Template for a Flask Web App ######
###### Requirements without Version Specifiers ######
# flask
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.0
attrs==24.2.0
blinker==1.8.2
boto3==1.35.24
botocore==1.35.24
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.4
colorama==0.4.6
dataclasses-json==0.6.7
faiss-cpu==1.8.0.post1
fastavro==1.9.7
filelock==3.16.1
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.0
langchain-cohere==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-experimental==0.3.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
MarkupSafe==2.1.5
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.7
packaging==24.1
pandas==2.2.3
parameterized==0.9.0
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pypdf==5.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.20.0
tqdm==4.66.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
